src/iris/requirements.txt
BabitMF
easyocr
numpy
opencv-python
Pillow
requests
datafog
presidio-image-redactor
wurlitzer
whisper-timestamped
torch